config/settings.py
"""
Settings Module - Configuration Management

This module loads and validates all configuration from .env file.
Uses python-dotenv to load environment variables and provides
a centralized Settings class for accessing configuration throughout the app.

Usage:
    from config.settings import Settings
    settings = Settings()
    token = settings.telegram_bot_token
"""

import os
from pathlib import Path
from typing import Optional
from dotenv import load_dotenv


class Settings:
    """
    Centralized configuration management class.

    Loads configuration from .env file and provides type-safe access
    to all configuration parameters with validation and defaults.
    """

    def __init__(self, env_file: str = ".env"):
        """
        Initialize settings from .env file.

        Args:
            env_file: Path to .env file (default: ".env" in project root)

        TODO:
        - Load .env file using load_dotenv()
        - Validate required environment variables exist
        - Raise clear errors if critical variables are missing
        - Set default values for optional variables
        """
        # TODO: Load environment variables
        load_dotenv(env_file)

        # TODO: Load and validate all settings
        self._load_telegram_settings()
        self._load_camera_settings()
        self._load_flask_settings()
        self._load_yolo_settings()
        self._load_motion_settings()
        self._load_logging_settings()
        self._load_system_settings()

    def _load_telegram_settings(self) -> None:
        """
        Load Telegram bot configuration.

        TODO:
        - Get TELEGRAM_BOT_TOKEN from environment
        - Get TELEGRAM_CHAT_ID from environment
        - Get ALERT_COOLDOWN with default 30
        - Validate token format (should be like: 123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11)
        - Validate chat_id is numeric
        """
        self.telegram_bot_token: str = os.getenv("TELEGRAM_BOT_TOKEN", "")
        self.telegram_chat_id: str = os.getenv("TELEGRAM_CHAT_ID", "")
        self.alert_cooldown: int = int(os.getenv("ALERT_COOLDOWN", "30"))

        # TODO: Add validation
        pass

    def _load_camera_settings(self) -> None:
        """
        Load camera configuration.

        TODO:
        - Get CAMERA_INDEX (default: 0)
        - Get CAMERA_WIDTH (default: 640)
        - Get CAMERA_HEIGHT (default: 480)
        - Get CAMERA_FPS (default: 15)
        - Validate all values are positive integers
        """
        self.camera_index: int = int(os.getenv("CAMERA_INDEX", "0"))
        self.camera_width: int = int(os.getenv("CAMERA_WIDTH", "640"))
        self.camera_height: int = int(os.getenv("CAMERA_HEIGHT", "480"))
        self.camera_fps: int = int(os.getenv("CAMERA_FPS", "15"))

        # TODO: Add validation
        pass

    def _load_flask_settings(self) -> None:
        """
        Load Flask server configuration.

        TODO:
        - Get FLASK_PORT (default: 5000)
        - Get FLASK_HOST (default: "0.0.0.0")
        - Get FLASK_DEBUG (default: False)
        - Validate port is in valid range (1024-65535)
        """
        self.flask_port: int = int(os.getenv("FLASK_PORT", "5000"))
        self.flask_host: str = os.getenv("FLASK_HOST", "0.0.0.0")
        self.flask_debug: bool = os.getenv("FLASK_DEBUG", "False").lower() == "true"

        # TODO: Add validation
        pass

    def _load_yolo_settings(self) -> None:
        """
        Load YOLO detection configuration.

        TODO:
        - Get YOLO_MODEL (default: "yolov5n.pt")
        - Get YOLO_CONFIDENCE (default: 0.6)
        - Get YOLO_IOU_THRESHOLD (default: 0.45)
        - Get YOLO_IMG_SIZE (default: 416)
        - Validate confidence and IOU are between 0.0 and 1.0
        - Validate img_size is positive and reasonable (e.g., 320-1280)
        """
        self.yolo_model: str = os.getenv("YOLO_MODEL", "yolov5n.pt")
        self.yolo_confidence: float = float(os.getenv("YOLO_CONFIDENCE", "0.6"))
        self.yolo_iou_threshold: float = float(os.getenv("YOLO_IOU_THRESHOLD", "0.45"))
        self.yolo_img_size: int = int(os.getenv("YOLO_IMG_SIZE", "416"))

        # TODO: Add validation
        pass

    def _load_motion_settings(self) -> None:
        """
        Load motion detection configuration.

        TODO:
        - Get MOTION_MIN_AREA (default: 500)
        - Get MOTION_FRAMES_THRESHOLD (default: 10)
        - Validate values are positive integers
        """
        self.motion_min_area: int = int(os.getenv("MOTION_MIN_AREA", "500"))
        self.motion_frames_threshold: int = int(os.getenv("MOTION_FRAMES_THRESHOLD", "10"))

        # TODO: Add validation
        pass

    def _load_logging_settings(self) -> None:
        """
        Load logging configuration.

        TODO:
        - Get LOG_LEVEL (default: "INFO")
        - Get LOG_FILE (default: "logs/security_system.log")
        - Get LOG_MAX_SIZE (default: 10 MB)
        - Get LOG_BACKUP_COUNT (default: 5)
        - Validate log level is valid (DEBUG, INFO, WARNING, ERROR, CRITICAL)
        """
        self.log_level: str = os.getenv("LOG_LEVEL", "INFO")
        self.log_file: str = os.getenv("LOG_FILE", "logs/security_system.log")
        self.log_max_size: int = int(os.getenv("LOG_MAX_SIZE", "10"))
        self.log_backup_count: int = int(os.getenv("LOG_BACKUP_COUNT", "5"))

        # TODO: Add validation
        pass

    def _load_system_settings(self) -> None:
        """
        Load general system configuration.

        TODO:
        - Get SYSTEM_NAME (default: "Smart Security System")
        - Get TIMEZONE (default: "Europe/Bucharest")
        - Get MAX_SNAPSHOTS (default: 100)
        - Get SNAPSHOT_RETENTION_DAYS (default: 7)
        """
        self.system_name: str = os.getenv("SYSTEM_NAME", "Smart Security System")
        self.timezone: str = os.getenv("TIMEZONE", "Europe/Bucharest")
        self.max_snapshots: int = int(os.getenv("MAX_SNAPSHOTS", "100"))
        self.snapshot_retention_days: int = int(os.getenv("SNAPSHOT_RETENTION_DAYS", "7"))

        # TODO: Add validation
        pass

    def validate(self) -> bool:
        """
        Validate all critical settings.

        Returns:
            bool: True if all settings are valid, False otherwise

        TODO:
        - Check Telegram credentials are set
        - Validate camera settings
        - Validate file paths exist or can be created
        - Validate GPIO pins are in valid range (for GPIOPins class)
        - Return True if valid, raise ValueError with clear message if invalid
        """
        # TODO: Implement comprehensive validation
        pass

    def __repr__(self) -> str:
        """String representation of settings (without sensitive data)."""
        return f"<Settings: {self.system_name}>"


# Create a singleton instance for easy import
settings = Settings()
