src/alerts/telegram_bot.py
"""
Telegram Bot Module - Remote Control and Alerts

This module provides Telegram bot interface for:
- System control (arm/disarm)
- Real-time alerts with photos
- Status queries
- Remote snapshots
- Event logs

Supported Commands:
- /start - Initialize bot
- /help - Show command list
- /arm - Arm the security system
- /disarm - Disarm the system
- /status - Get system status
- /snapshot - Get current camera frame
- /logs - View recent events

Usage:
    from src.alerts.telegram_bot import TelegramBot

    bot = TelegramBot()
    bot.start()

    # Send alert
    bot.send_alert("CRITICAL: Person detected!", frame=image, alert_type="person")

    # Bot runs until stopped
    bot.stop()
"""

from telegram import Update, Bot
from telegram.ext import Application, CommandHandler, ContextTypes
import io
from PIL import Image
import numpy as np
from typing import Optional, Callable
from datetime import datetime

from config.settings import settings


class TelegramBot:
    """
    Telegram bot for remote control and notifications.

    Handles commands from authorized users and sends alerts with images.
    """

    def __init__(self, token: Optional[str] = None, chat_id: Optional[str] = None):
        """
        Initialize Telegram bot.

        Args:
            token: Bot token from BotFather (default from settings)
            chat_id: Authorized chat ID (default from settings)

        TODO:
        - Load token and chat_id from settings if not provided
        - Initialize bot application
        - Register command handlers
        - Initialize callback storage for system control
        """
        self.token = token or settings.telegram_bot_token
        self.chat_id = chat_id or settings.telegram_chat_id

        self.application = None
        self.bot = None

        # Callbacks for system control (will be set by SystemManager)
        self.on_arm_callback = None
        self.on_disarm_callback = None
        self.get_status_callback = None
        self.get_snapshot_callback = None

    def start(self) -> None:
        """
        Start Telegram bot with polling.

        TODO:
        - Create Application: Application.builder().token(self.token).build()
        - Register command handlers:
            - /start -> self.cmd_start
            - /help -> self.cmd_help
            - /arm -> self.cmd_arm
            - /disarm -> self.cmd_disarm
            - /status -> self.cmd_status
            - /snapshot -> self.cmd_snapshot
            - /logs -> self.cmd_logs
        - Start polling in background
        - Send startup message to chat
        """
        # TODO: Implement bot startup
        pass

    async def cmd_start(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /start command.

        TODO:
        - Check if user is authorized (compare chat_id)
        - Send welcome message with available commands
        - If unauthorized, send rejection message
        """
        # TODO: Implement /start command
        pass

    async def cmd_help(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /help command.

        TODO:
        - Send formatted help message with all commands
        - Include brief description of each command
        - Use markdown formatting for readability
        """
        # TODO: Implement /help command
        pass

    async def cmd_arm(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /arm command.

        TODO:
        - Check authorization
        - Call on_arm_callback if set
        - Send confirmation message
        - Include current system status
        """
        # TODO: Implement /arm command
        pass

    async def cmd_disarm(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /disarm command.

        TODO:
        - Check authorization
        - Call on_disarm_callback if set
        - Send confirmation message
        """
        # TODO: Implement /disarm command
        pass

    async def cmd_status(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /status command.

        TODO:
        - Call get_status_callback to get system status
        - Format status message:
            - System state (armed/disarmed)
            - Uptime
            - CPU/RAM usage
            - Last detection time
            - Statistics (detections today, etc.)
        - Send formatted message
        """
        # TODO: Implement /status command
        pass

    async def cmd_snapshot(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /snapshot command.

        TODO:
        - Call get_snapshot_callback to get current frame
        - Convert frame to image
        - Send photo to user
        - Include timestamp in caption
        """
        # TODO: Implement /snapshot command
        pass

    async def cmd_logs(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """
        Handle /logs command.

        TODO:
        - Read recent log entries (last 10-20)
        - Format as readable message
        - Send to user
        - Consider pagination for many logs
        """
        # TODO: Implement /logs command
        pass

    def send_alert(
        self,
        message: str,
        frame: Optional[np.ndarray] = None,
        alert_type: str = "info"
    ) -> None:
        """
        Send alert message with optional image.

        Args:
            message: Alert message text
            frame: Optional frame to send as photo
            alert_type: Alert type (info, warning, critical)

        TODO:
        - Format message with emoji based on alert_type:
            - info: ℹ️
            - warning: ⚠️
            - critical: 🚨
        - Add timestamp
        - If frame provided:
            - Convert numpy array to PIL Image
            - Send as photo with caption
        - Else:
            - Send as text message
        - Handle send errors gracefully
        """
        # TODO: Implement alert sending
        pass

    def send_message(self, text: str) -> None:
        """
        Send plain text message.

        TODO:
        - Use bot.send_message()
        - Send to authorized chat_id
        - Handle errors
        """
        # TODO: Implement message sending
        pass

    def send_photo(self, frame: np.ndarray, caption: str = "") -> None:
        """
        Send photo with optional caption.

        Args:
            frame: Image frame (numpy array)
            caption: Photo caption

        TODO:
        - Convert numpy array to bytes
        - Use bot.send_photo()
        - Include caption if provided
        - Handle errors
        """
        # TODO: Implement photo sending
        pass

    def _is_authorized(self, chat_id: int) -> bool:
        """
        Check if user is authorized.

        TODO:
        - Compare chat_id with self.chat_id
        - Return True if authorized
        - Could extend to support multiple authorized users
        """
        # TODO: Implement authorization check
        pass

    def register_callbacks(
        self,
        on_arm: Optional[Callable] = None,
        on_disarm: Optional[Callable] = None,
        get_status: Optional[Callable] = None,
        get_snapshot: Optional[Callable] = None,
    ) -> None:
        """
        Register callback functions for system control.

        TODO:
        - Store callback functions
        - Will be called by command handlers
        - Allows SystemManager to control system via Telegram
        """
        self.on_arm_callback = on_arm
        self.on_disarm_callback = on_disarm
        self.get_status_callback = get_status
        self.get_snapshot_callback = get_snapshot

    def stop(self) -> None:
        """
        Stop Telegram bot.

        TODO:
        - Stop polling
        - Send shutdown message
        - Cleanup resources
        """
        # TODO: Implement bot shutdown
        pass

    def __repr__(self) -> str:
        """String representation."""
        return f"<TelegramBot: chat_id={self.chat_id}>"


if __name__ == "__main__":
    """Test Telegram bot."""
    print("Telegram Bot test - TODO: Implement test code")
    pass
